Move environment_canada coordinator to separate module
